third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-transition-transform-list-expected.txt

This is a testharness.js-based test.
FAIL A custom property of type <transform-list> can yield a CSS Transition assert_equals: A single animation is running expected 1 but got 0
Harness: the test ran to completion.
